requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
python=3.8
numpy=1.23
scikit-learn=1.2
pytorch=1.12
